chore: use only types for accounts
